Apps Code/master/backend/requirements.txt
Django==2.2.13
djongo==1.3.0
django-pipeline
django-tabbed-admin
Pillow 
django-tinymce4-lite
django-clear-cache 
requests
pygal
pycountry
python-dateutil
watchdog 
sqlparse==0.2.4
djangorestframework==3.11.0
django-cors-headers
rest-meets-djongo